Move imperialprobedroid to droid package
