+++ src/superdopesquad/superdopejedimod/entity/droid/ImperialProbeDroidModel.java
@@ -1,4 +1,4 @@
-package superdopesquad.superdopejedimod.entity;
+package superdopesquad.superdopejedimod.entity.droid;
 
 import java.util.ArrayList;
 import java.util.List;
